requirements.txt
langchain==0.1.0
langchain-community==0.0.10
pandas==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.0